requirements.txt
Django>=4.2
djangorestframework>=3.14
django-filter>=23.5
drf-spectacular>=0.27
djangorestframework-simplejwt
python-decouple